Fix SDK with latest dev changes.
This commit fixes the SDK based on the latest changes that were made to the
`dev` branch.
