+++ azure-functions-sdk/src/commands/new.rs
@@ -14,22 +14,26 @@ use std::{
     io::Read,
     path::Path,
 };
-use syn::{parse::Parser, parse_file, punctuated::Punctuated, Ident, Item, Token};
+use syn::{self, parse::Parser, parse_file, punctuated::Punctuated, Item, Token};
 
+mod activity;
 mod blob;
 mod cosmos_db;
 mod event_grid;
 mod event_hub;
 mod http;
+mod orchestration;
 mod queue;
 mod service_bus;
 mod timer;
 
+pub use self::activity::Activity;
 pub use self::blob::Blob;
 pub use self::cosmos_db::CosmosDb;
 pub use self::event_grid::EventGrid;
 pub use self::event_hub::EventHub;
 pub use self::http::Http;
+pub use self::orchestration::Orchestration;
 pub use self::queue::Queue;
 pub use self::service_bus::ServiceBus;
 pub use self::timer::Timer;
@@ -103,6 +107,28 @@ fn create_function(name: &str, template: &str, data: &Value, quiet: bool) -> Res
     Ok(())
 }
 
+fn format_path(path: &syn::Path) -> String {
+    use std::fmt::Write;
+
+    let mut formatted = String::new();
+    if path.leading_colon.is_some() {
+        formatted.push_str("::");
+    }
+
+    let mut first = true;
+    for segment in &path.segments {
+        if first {
+            first = false;
+        } else {
+            formatted.push_str("::");
+        }
+
+        write!(formatted, "{}", segment.ident).unwrap();
+    }
+
+    formatted
+}
+
 fn export_function(name: &str) -> Result<(), String> {
     let mut file =
         File::open("src/functions/mod.rs").map_err(|_| "'src/functions/mod.rs' does not exist.")?;
@@ -124,11 +150,11 @@ fn export_function(name: &str) -> Result<(), String> {
             Item::Macro(m) => {
                 if last_segment_in_path(&m.mac.path).ident == "export" {
                     exports.extend(
-                        Punctuated::<Ident, Token![,]>::parse_terminated
+                        Punctuated::<syn::Path, Token![,]>::parse_terminated
                             .parse2(m.mac.tokens)
-                            .map_err(|_| "failed to parse 'export!' macro.")?
+                            .map_err(|e| format!("failed to parse 'export!' macro: {}", e))?
                             .into_iter()
-                            .map(|i| i.to_string()),
+                            .map(|p| format_path(&p)),
                     );
                 }
             }
@@ -139,7 +165,7 @@ fn export_function(name: &str) -> Result<(), String> {
     modules.push(name.to_string());
     modules.sort();
 
-    exports.push(name.to_string());
+    exports.push(format!("{}::{}", name, name));
     exports.sort();
 
     create_from_template(
@@ -225,76 +251,6 @@ impl<'a> From<&'a ArgMatches<'a>> for New<'a> {
             quiet: args.is_present("quiet"),
             color: args.value_of("color"),
             args,
-        }
-    }
-}
-
-struct Activity<'a> {
-    name: &'a str,
-}
-
-impl<'a> Activity<'a> {
-    pub fn create_subcommand<'b>() -> App<'a, 'b> {
-        SubCommand::with_name("activity")
-            .about("Creates a new Activity Function for Durable Functions.")
-            .arg(
-                Arg::with_name("name")
-                    .long("name")
-                    .short("n")
-                    .value_name("NAME")
-                    .help("The name of the new Azure Function.")
-                    .required(true),
-            )
-    }
-
-    pub fn execute(&self, quiet: bool) -> Result<(), String> {
-        let data = json!({
-            "name": self.name,
-        });
-
-        create_function(self.name, "activity.rs", &data, quiet)
-    }
-}
-
-impl<'a> From<&'a ArgMatches<'a>> for Activity<'a> {
-    fn from(args: &'a ArgMatches<'a>) -> Self {
-        Activity {
-            name: args.value_of("name").unwrap(),
-        }
-    }
-}
-
-struct Orchestration<'a> {
-    name: &'a str,
-}
-
-impl<'a> Orchestration<'a> {
-    pub fn create_subcommand<'b>() -> App<'a, 'b> {
-        SubCommand::with_name("orchestration")
-            .about("Creates a new Orchestration Function for Durable Functions.")
-            .arg(
-                Arg::with_name("name")
-                    .long("name")
-                    .short("n")
-                    .value_name("NAME")
-                    .help("The name of the new Azure Function.")
-                    .required(true),
-            )
-    }
-
-    pub fn execute(&self, quiet: bool) -> Result<(), String> {
-        let data = json!({
-            "name": self.name,
-        });
-
-        create_function(self.name, "orchestration.rs", &data, quiet)
-    }
-}
-
-impl<'a> From<&'a ArgMatches<'a>> for Orchestration<'a> {
-    fn from(args: &'a ArgMatches<'a>) -> Self {
-        Orchestration {
-            name: args.value_of("name").unwrap(),
         }
     }
 }
