+++ azure-functions-sdk/src/templates/new-app/functions_mod.rs.hbs
@@ -6,6 +6,6 @@
 {{/if~}}
 // Export the Azure Functions here.
 azure_functions::export! { {{~#each exports}}
-    {{this}}::{{this}},{{else~}}{{/each~}}{{~#if exports}}
+    {{this}},{{else~}}{{/each~}}{{~#if exports}}
 {{/if~}}
 }
